requirements.txt
pandas==2.2.0
numpy==1.26.0
scikit-learn==1.4.0
lightgbm==4.3.0
tensorflow==2.15.0
shap==0.44.0
evidently==0.4.16
scipy==1.12.0
statsmodels==0.14.1
streamlit>=1.32.0
packaging>=23.0
plotly==5.19.0
google-generativeai==0.5.0
python-dotenv==1.0.0
dbt-core==1.7.0
dbt-sqlite==1.7.0
pytest==8.0.0
joblib==1.3.2
openpyxl==3.1.2
xlrd==2.0.1
